makes install routine more failure save
this reduces the need to sleep before action after nodes are created from 30 to 5 seconds. The issue is, that the new nodes are randomly failing during `apt` commands the first 30 - 60 seconds after creation, despite they are created in 10 seconds.

The SSH client also behaves more silently in failures, so the progress bars are not bugging.

fix #106